create config dir | move config | fix config reflect
