+++ Gems/FPSProfiler/Code/fpsprofiler_editor_private_files.cmake
@@ -1,7 +1,5 @@
 
 set(FILES
-        Source/Tools/FPSProfilerConfig.cpp
-        Source/Tools/FPSProfilerConfig.h
     Source/Tools/FPSProfilerEditorSystemComponent.cpp
     Source/Tools/FPSProfilerEditorSystemComponent.h
 )
