+++ Gems/FPSProfiler/Code/fpsprofiler_private_files.cmake
@@ -4,4 +4,6 @@ set(FILES
     Source/FPSProfilerModuleInterface.h
     Source/Clients/FPSProfilerSystemComponent.cpp
     Source/Clients/FPSProfilerSystemComponent.h
+    Source/Configurations/FPSProfilerConfig.cpp
+    Source/Configurations/FPSProfilerConfig.h
 )
